Update to fix PHP errors with regional tabs
When passing [tabs] to the paraneue_dosomething_edit_regional_admin_tabs() function, if the #primary item in the array is empty, then the following code in the function shouldn't execute. It previously did and PHP would throw warnings. So now we check to make sure it's not and empty array; if it is, then we return back with the untouched .