Wrapping up fix for translated Facts.
To minimize the potential for breaking lots of campaign related code, took a more subtle approach to fixing the translation bug related to Facts. The EntityDrupalWrapper class was causing problems with grabbing the correct facts related to regional content. Instead, created a separate function to handle grabbing that data from the node, without interrupting how the rest of the campaign node data is loaded when constructing the campaign object. It allows for grabbing content if there are translations and ones specific to the language defined by path (/us/ or /mx/ etc) or just grabbing the available content if no language specific options available, but there is definitely content in the field.