I feel dumb.
The  needed to be set inside the foreach loop so that it would be reset each time it looped for each fact (both problem and solution facts). With it being set outside the loop, if the fact problem existed the  would be set for that fact, and then on the second run, if the solution fact was missing, it would keep the  set as the one for the problem fact. Shiza.