New function added to get entity if not already laoded.
As it turns out not all stages in Drupal request/page data load cycle are created equal. During the loading of the node for the reportback share, when the node is loaded for the campaign, not all the nested nodes (ie: facts) are loaded as well; instead those nodes are loaded with just a target_id indicator. So sometimes we need to roll up our sleeves and load the node ourselves