Setting proper checks to not have undefined properties.
If you are an anonymous requester retrieving data from the api, some properies such as 'flagged' and 'why_participated' are not exposed, but were not being checked properly to provide a sensible default. Using the dosomething_helpers_isset() function allows us to check the data for the property and if not set assign a default in one fell swoop.