Refactoring API RB item status security.
The logic to remove restricted statuses has been switched around from being in the dosomething_reportback_item.query.inc file to being a private method of the ReportbackItemTransformer class. The security measures in place are really only being used by the API, and this also allows for securing the filters earlier on in the transformation process, rather than later when it hits the query builder, allowing the need to process the filters only once instead of multiple times. These changes were also required for how the totals are calculated and relying on accurate use of the RB item statuses.