Updating show method for RB Item results with some security.
While the index collection of RB Items from the API is now doing a better job of securing content that should not be accessible to anonymous API requests, the show view for a single RB Item was not. This update adds a method to check whether an RB Item returned via requests to  should be accessible or not, and if the anonymous user does not have access to flagged, excluded, and pending items that are specifically requested by their ID, then an exception is thrown and caught via the normal process within the API request cycle.