Fix for missing CTAs on Fact pages.
At some point the expected ['content'] array data changed and no longer houses the CTA information (along with Fact page images), so the Fact page template was breaking on both CTA content and image content. The solution involved adding a bit of new logic to account for  information as well as adjusting where the data was being retrieved from. It now refers to the data in the main  array.