Updating the templating process for 'photo' pattern.
Using the ReportbackItems class for retrieving gallery items means that the structure of the data has changed from being an array to now an object with a different set of properties, etc. This updates the templating process to account for the new data structure that gets passed to the template.