Necessary updates for accurate RB Item tallying.
Turns out the old methods we were using for tallying totals of promoted and approved RB items for specified campaigns was no longer suitable. We needed a more robust and accurate tallying system which can account for nuances in what parameters are passed either via the API or when querying the ReportbackItem class. We kept getting inaccurate totals because the old methods only queried based on a status, but now the ReportbackItem query system supports many other parameters that affect the tallied results. This commit updates the entire approach to tallying RB Items and also begins process of caching the results in Redis using the ApiCache class.