Holy code refactor Batman\!
So turns out a few of the old functions use for calculating totals for reportback items and with that the generation of the pagination object for the API was all kinds of messed up. The prior commit helped with starting to set things on the right path for tallying totals as well as caching them, but that was only for the DS app side of things. This commit helps both the DS app AND the API play nice. Tallying the totals is now entirely based on the extent of what filters are used for a query, and the pagination object uses this now correct total value to figure out the pagination as well as now fully incorporating all url params properly when generating the prev/next urls.