Fixes broken endpoint when no parameters passed.
Turns out I had moved the  variable generation inside a scoped logic block; if the logic didn't pass then the variable wasn't set and it's being used a little farther down in the code so it needs to be set regardless. Moved it up and out of the logic block. Works fine now.