Adding new getter methods and new properties.
Adding new properties for language and translations as well as a property to contain campaign runs data for each parent campaign. This also adds a couple handy getter methods for the core drupal node and variables data, so they can be retrieved since they are protected properties that will not otherwise output.