Update to fix up Campaign::get() to return logical data.
This is a quick update to the returned data from Campaign::get() static method. It allows passing single string or multiple ids as an array, however if passing a single string it makes it a bit cumbersome to remember to pluck the first item out as the actual data you want to use. This update fixes that so if you pass a single string id, it'll return the Campaign object, if you pass an array of ids it'll return a collection of campaign objects.