Update to fix Reportback::get() static method.
This method used to return a collection of Reportbacks in an array even if only one Reportback was requested. This is because you can request a series of Reportbacks if you pass an array of RB ids. Now when requesting 1 RB the method knows this and returns just the single RB which is much cleaner and easier for other developers. This commit also does a little bit of cleanup in other classes.