Updates to RB save function and assoc functions.
There is more than one location that uses the dosomething_reportback_save() function, so we need to account for them all. This update begins the setup to DRY the code needed to update all the functions, by handling the user, campaign and run id in one location instead of various potential locations.