Update so now the ApiCache can be cleared when a general cache clear is performed.
his enables the use of the 'CACHE_TEMPORARY' setting, so that the data stays in cache until a general cache clear is performed via the Admin interface or through Drush, while still allowing the ability to customize and pass a custom number of hours to cache the data for or to cache it permanently. The default it to cache it temporarily.