Updates to add ApiClient class for API calls in JS.
We can now use this ApiClient class to handle calls to the Phoenix API and in the future the rest of our services (after a few upgrades). It updates the FeatureCard to start using the new class as well as starts the implementation for use of the class as well in the ReportbackItemKudos component.