Adding new affiliate configurations form in CMS specific to each campaign.
This adds a custom form for specifying affiliate information on a per campaign basis and updates the pitch page to render the form with the custom settings as needed, when a user is signed in. It also adds a new per campaign export form where it will allow export of the users that opted in (opting out for this first one ;), but could be used for exporting other campaign specific data in the future.