Issue 7376: Upgrade Drupal Core from 7.51 to 7.54
We are a few versions behind the recommended version of Drupal 7 core. One of the releases includes a security release.
We are currently on version 7.51.
The following lists the changes in the newer versions:

**7.52:**
Maintenance and security release of the Drupal 7 series. ([Release notes](https://www.drupal.org/project/drupal/releases/7.52))

Fixes:

- Inconsistent name for term access query _(Less critical)_
- Confirmation forms allow external URLs to be injected _(Moderately critical)_

**7.53:**

Maintenance release that addresses the regression with drag-and-drop behavior that was introduced in Drupal 7.51 (and persisted in Drupal 7.52) for certain web browsers and certain jQuery versions. ([Release notes](https://www.drupal.org/project/drupal/releases/7.53))

**7.54:**

Maintenance release of the Drupal 7 series. Includes a variety of improvements and bug fixes (no major, non-backwards-compatible new functionality).  ([Release notes](https://www.drupal.org/project/drupal/releases/7.534))

Three change records are associated with the 7.54 release:

- [Modules can now define theme engines](https://www.drupal.org/node/2826480)
- [Drupal 7 adds menu tree render structure to (pre-)process hooks for theme_menu_tree()](https://www.drupal.org/node/2827134)
- [New drupal_is_https() API function added](https://www.drupal.org/node/2824590)

We should keep Drupal core up to date, _especially_ when there are security releases.