No Ticket: Remove User from metatag configuration
This is related to work done for issue #: 7392 - Upgrade Metatag module from 7.x-1.4 to 7.x-1.21

On Thor, with the larger dataset, the upgrade process took 1 day and 21 hours. The reason for this appears to be that metatags
is updating all of the user entities (there's a lot of them). This is completely unnecessary because user profiles are not
exposed to the public and thus having metatags on them is also unnecessary.

This commit removes User from metatag configuration. Also includes a re-export (minor change) of the twitter image metatag.